requirements.txt
Byaldi==0.0.4
colpali_engine==0.3.0
langchain==0.3.1
langchain_community==0.3.1
langchain_core==0.3.6
langchain_openai==0.2.0
pdf2image==1.17.0
Pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
twilio==9.3.2
wikipedia==1.4.0
wolframalpha==5.1.3

deepgram-sdk==3.*
websocket==0.2.1
elevenlabs==1.9.0
pydub==0.25.1
langchain-groq

fastapi~=0.115.0
langchain-core~=0.3.6
langchain-community~=0.3.1
langchain-openai~=0.2.0
starlette~=0.38.6
requests~=2.32.3